Updating UI to work with backend logic

+++ src/main/java/dev/andepark/minicasino/models/Game.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 public class Game {
     private int id;
     private String name;
-    private int chanceOfWinning;
+    private double chanceOfWinning;
     private double multiplier;
     private double maxBet;
     private double minBet;
